separate shared logic from boxlite implementation
